# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/setup:
    get:
      operationId: getOnboardingSetup
      tags: [m365]
      summary: Get stored onboarding setup details with secrets redacted
      responses:
        "200":
          description: Redacted onboarding setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSetup"
    patch:
      operationId: patchOnboardingSetup
      tags: [m365]
      summary: Save onboarding setup details securely
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingSetupPatch"
      responses:
        "200":
          description: Updated redacted onboarding setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSetup"
...
components:
...
  schemas:
...
    OnboardingSetup:
      type: object
      properties:
        tenantId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        clientSecret:
          type: string
          nullable: true
        hasClientSecret:
          type: boolean
        setupComplete:
          type: boolean
        setupCompletedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - tenantId
        - clientId
        - clientSecret
        - hasClientSecret
        - setupComplete
        - setupCompletedAt
        - createdAt
        - updatedAt
    OnboardingSetupPatch:
      type: object
      properties:
        tenantId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        clientSecret:
          type: string
          nullable: true
        setupComplete:
          type: boolean
        acknowledgedMissingPermissions:
          type: array
          items:
            type: string
          nullable: true
...